A few changles...
I added a method to calculate the return angle of the ball based upon where it strikes the paddle.

I also made it so that the ball gradually speeds up as the game progresses.

I renamed a couple of variables (e.g. lastPaddle is now lastBounce because it now also keeps track of when it bounces off of the top or the bottom.